backend/requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.2
djangorestframework==3.14.0
idna==3.4
kombu==5.2.4
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
python-dotenv==1.0.0
pytz==2023.3
redis==4.5.4
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
